feat: deno?
Basically: can we rebuild the parts of webpack that we need with deno? So far, the answer is: yes, but we're not going to get much help doing it.

This code is building and serving a javascript bundle (from `wasm-pack build --target web`) alongside our index.html, but it's failing to load the `wasm` bits before trying to use them.